agent_apps/openai_exchange_rate_agent/requirements.txt
openai-agents
streamlit
